bounding_boxes/requirements.txt
Django==5.1.3
Pillow==11.0.0
gunicorn==23.0.0
numpy==2.1.3